Move useGetMostRecentOpenConversation to hooks and simplify conversation sorting
