+++ packages/help-center/src/hooks/use-get-most-recent-open-conversation.tsx
@@ -1,0 +1,50 @@
+import { HelpCenterSelect } from '@automattic/data-stores';
+import { useGetSupportInteractions } from '@automattic/odie-client/src/data';
+import { useSelect } from '@wordpress/data';
+import {
+	getZendeskConversations,
+	getConversationsFromSupportInteractions,
+	sortConversationByMostRecent,
+} from '../components/utils';
+import { HELP_CENTER_STORE } from '../stores';
+
+export const useGetMostRecentOpenConversation = () => {
+	let userHasRecentOpenConversation = false;
+	let supportInteractionId = null;
+
+	const { isChatLoaded } = useSelect( ( select ) => {
+		const store = select( HELP_CENTER_STORE ) as HelpCenterSelect;
+		return {
+			isChatLoaded: store.getIsChatLoaded(),
+		};
+	}, [] );
+
+	const { data: supportInteractionsResolved, isLoading: isLoadingResolvedInteractions } =
+		useGetSupportInteractions( 'zendesk', 100, 'resolved' );
+
+	const { data: supportInteractionsOpen, isLoading: isLoadingOpenInteractions } =
+		useGetSupportInteractions( 'zendesk', 10, 'open' );
+
+	const isLoadingInteractions = isLoadingResolvedInteractions || isLoadingOpenInteractions;
+
+	if ( isChatLoaded && getZendeskConversations && ! isLoadingInteractions ) {
+		const allConversations = getZendeskConversations();
+		const supportInteractions = [
+			...( supportInteractionsResolved || [] ),
+			...( supportInteractionsOpen || [] ),
+		];
+
+		const filteredConversations = getConversationsFromSupportInteractions(
+			allConversations,
+			supportInteractions
+		);
+
+		const recentConversations = sortConversationByMostRecent( filteredConversations );
+
+		if ( recentConversations?.length > 0 ) {
+			supportInteractionId = recentConversations[ 0 ]?.metadata?.supportInteractionId;
+			userHasRecentOpenConversation = true;
+		}
+	}
+	return { userHasRecentOpenConversation, supportInteractionId };
+};
